Add and Use Anton Font for Title
Adds the Anton font and uses it for the title screen.

To help preserve the licenses of external resources I did not create, such as
the SIL Open Font License 1.1 for Anton, these resources will go in
external_assets with their licenses. I have included a notice to this effect.